Extracted out `fca utils` out of fca import
