+++ config/initializers/slack.rb
@@ -1,0 +1,6 @@
+require 'slack-ruby-client'
+
+Slack.configure do |config|
+  config.token = ENV['SLACK_API_TOKEN']
+  fail 'Missing ENV[SLACK_API_TOKEN]!' unless config.token
+end
