+++ lib/fca.rb
@@ -3,6 +3,7 @@ module FCA
 end
 
 require_relative 'fca/config'
+require_relative 'fca/utils'
 require_relative 'fca/import'
 require_relative 'fca/row'
 require_relative 'fca/query'
