+++ lib/fca/config.rb
@@ -3,13 +3,10 @@ require 'forwardable'
 
 module FCA
   class Config
-    attr_writer :log_file, :log_level
-    attr_reader :emails
-
     class << self
       extend Forwardable
 
-      def_delegators :config, :logger, :emails
+      def_delegators :config, :logger, :notify, :hostname
 
       def configure
         yield config if block_given?
@@ -21,6 +18,9 @@ module FCA
     end
 
     LOG_LEVEL = %w(UNKNOWN FATAL ERROR WARN INFO DEBUG).freeze
+
+    attr_accessor :notify, :hostname
+    attr_writer :log_file, :log_level
 
     def logger
       @logger ||= ::Logger.new(log_file, ::File::APPEND)
@@ -36,10 +36,6 @@ module FCA
     def log_level
       level = LOG_LEVEL.detect { |level| level == @log_level.to_s.upcase.strip } || 'INFO'
       Object.const_get("Logger::#{level}")
-    end
-
-    def emails=(s)
-      @emails = s.split(',') if s
     end
   end
 end
